backend/config.py
import os

# محاولة تحميل dotenv إن وجد (بدون كسر)
try:
    from dotenv import load_dotenv
    BASE_DIR = os.path.dirname(os.path.abspath(__file__))
    load_dotenv(os.path.join(BASE_DIR, ".env"))
except Exception:
    BASE_DIR = os.path.dirname(os.path.abspath(__file__))


class Config:
    SECRET_KEY = os.environ.get("SECRET_KEY", "dev-secret-key")

    SQLALCHEMY_DATABASE_URI = os.environ.get(
        "DATABASE_URL",
        "sqlite:///" + os.path.join(BASE_DIR, "maintenance.db").replace("\\", "/")
    )

    SQLALCHEMY_TRACK_MODIFICATIONS = False
