backend/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
reportlab==4.2.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
argostranslate
python-dotenv
